refactor: Organize code to remove flake8 issues
