+++ archivebox/cli/archivebox_config.py
@@ -8,7 +8,8 @@ import argparse
 
 from typing import Optional, List, IO
 
-from ..main import config, docstring
+from ..main import config
+from ..util import docstring
 from ..config import OUTPUT_DIR
 from ..logging_util import SmartFormatter, accept_stdin
