+++ archivebox/core/welcome_message.py
@@ -2,5 +2,4 @@ from archivebox.logging_util import log_shell_welcome_msg
 
 
 if __name__ == '__main__':
-    from main import *
     log_shell_welcome_msg()
